ZCS-1825: register as listener with ZMailbox as soon as ZMailbox is created
