+++ store/src/java/com/zimbra/cs/imap/ImapRemoteSession.java
@@ -19,29 +19,15 @@ package com.zimbra.cs.imap;
 import java.util.TreeMap;
 
 import com.zimbra.client.ZMailbox;
-import com.zimbra.client.event.ZEventHandler;
 import com.zimbra.common.mailbox.BaseItemInfo;
 import com.zimbra.common.mailbox.MailItemType;
-import com.zimbra.common.mailbox.MailboxStore;
 import com.zimbra.common.service.ServiceException;
 import com.zimbra.common.util.ZimbraLog;
 import com.zimbra.cs.session.PendingModifications;
 import com.zimbra.cs.session.PendingModifications.Change;
 import com.zimbra.cs.session.PendingRemoteModifications;
-import com.zimbra.soap.type.AccountWithModifications;
 
 public class ImapRemoteSession extends ImapListener {
-    private final ZEventHandler zMailboxEventHandler = new ZEventHandler() {
-        @Override
-        public void handlePendingModification(int changeId, AccountWithModifications info) throws ServiceException {
-            ZimbraLog.imap.debug("Handling modification from ZMailbox");
-            MailboxStore store = getMailbox();
-            if(store != null && store instanceof ZMailbox) {
-                ImapServerListenerPool.getInstance().get((ZMailbox)store).notifyAccountChange(info);
-            }
-        }
-    };
-
     protected class PagedRemoteFolderData extends ImapListener.PagedFolderData {
 
         PagedRemoteFolderData(String cachekey, ImapFolder i4folder) {
@@ -92,9 +78,6 @@ public class ImapRemoteSession extends ImapListener {
     protected ImapRemoteSession(ImapMailboxStore imapStore, ImapFolder i4folder, ImapHandler handler) throws ServiceException {
         super(imapStore, i4folder, handler);
         mailbox = imapStore.getMailboxStore();
-        if(mailbox instanceof ZMailbox) {
-            ((ZMailbox)mailbox).addEventHandler(zMailboxEventHandler);
-        }
     }
 
     @Override
